+++ packages/shadertoy_cli/lib/shadertoy_cli.dart
@@ -1,6 +1,4 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
+/// A command line tool to interact with the shadertoy rest and site apis
 library shadertoy_cli;
 
 export 'src/cli.dart';