+++ packages/shadertoy_cli/lib/src/cli.dart
@@ -1,5 +1,3 @@
-import 'dart:async';
-
 import 'package:args/command_runner.dart';
 
 import 'command/comments.dart';
@@ -9,8 +7,11 @@ import 'command/shader.dart';
 import 'command/sqlite.dart';
 import 'command/user.dart';
 
+/// Main entry point to execute the command line shadertoy client
+///
+/// * [args]: The arguments provided to the command line tool
 Future run(List<String> args) async {
-  final runner = CommandRunner('shadertoy', 'Command-line shadertoy client')
+  final runner = CommandRunner('shadertoy', 'Command line shadertoy client')
     ..addCommand(UserCommand())
     ..addCommand(ShaderCommand())
     ..addCommand(CommentsCommand())