+++ packages/shadertoy_cli/lib/src/command/multi.dart
@@ -5,18 +5,17 @@ import 'package:shadertoy/shadertoy_api.dart';
 import 'package:shadertoy_client/shadertoy_client.dart';
 import 'package:shadertoy_sqlite/shadertoy_sqlite.dart';
 
+import 'api.dart';
 import 'base.dart';
 
-abstract class MultiCommand extends BaseCommand
+/// Provides a base definition of a command that interacts with a
+/// shadertoy http or store client
+abstract class MultiCommand extends APICommand
     implements ExtendedClientCommand {
+  /// Builds an [MultiCommand]
   MultiCommand() {
-    argParser
-      ..addOption('user', abbr: 'u', help: 'The user', valueHelp: 'user')
-      ..addOption('password',
-          abbr: 'p', help: 'The password', valueHelp: 'password')
-      ..addOption('apiKey', abbr: 'k', help: 'The api key', valueHelp: 'apiKey')
-      ..addOption('file',
-          abbr: 'f', help: 'The database location', valueHelp: 'file');
+    argParser.addOption('file',
+        abbr: 'f', help: 'The database location', valueHelp: 'file');
   }
 
   @override