+++ packages/shadertoy_cli/lib/src/command/sqlite.dart
@@ -6,12 +6,14 @@ import 'package:shadertoy_sqlite/shadertoy_sqlite.dart';
 
 import 'database.dart';
 
+/// Command to scrape shadertoy data from the rest and site API into a SQLite database
 class SqliteCommand extends DatabaseCommand {
   @override
   final name = 'sqlite';
   @override
   final description = 'Scrapes shaders to a sqlite database';
 
+  /// Builds a [SqliteCommand]
   SqliteCommand();
 
   @override