+++ packages/shadertoy_client/example/site/find_shader_by_id.dart
@@ -4,7 +4,7 @@ import 'package:shadertoy_client/shadertoy_client.dart';
 void main(List<String> arguments) async {
   final client = newShadertoySiteClient();
 
-  final result = await client.findShaderById('Mt3XW8');
+  final result = await client.findShaderById('WltXD8');
   print(result.shader?.info.id);
   print('\tName: ${result.shader?.info.name}');
   print('\tUserName: ${result.shader?.info.userId}');