+++ packages/shadertoy_client/example/ws/find_shader_byid.dart
@@ -16,7 +16,7 @@ void main(List<String> arguments) async {
   }
 
   final client = newShadertoyWSClient(apiKey);
-  final result = await client.findShaderById('Mt3XW8');
+  final result = await client.findShaderById('WltXD8');
   print(result.shader?.info.id);
   print('\tName: ${result.shader?.info.name}');
   print('\tUserName: ${result.shader?.info.userId}');