+++ packages/shadertoy_sqlite/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.2
+
+- Updated dependencies
+
 ## 2.0.1
 
 - Corrected broken links in the README file